+++ history.md
@@ -2,6 +2,7 @@
 
 - Support code lens feature on vim9.
 - `codeLens.subseparator` default changed to `|`, like VSCode.
+- Add configuration `coc.preferences.enableGFMBreaksInMarkdownDocument`, default to `true`
 
 # 2022-11-19